separated spec files for separator
