+++ cypress/integration/v2/blocks/separator/other-tests.spec.js
@@ -1,0 +1,19 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import {
+	blockExist,
+	blockError,
+	switchLayout,
+} from '.'
+
+describe( 'Separator Block ( Other Tests )', registerTests( [
+	blockExist,
+	blockError,
+	switchLayout,
+] ) )
